requirements.txt
google-genai>=0.3.0
google-adk>=0.1.0
streamlit>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
